requirements.txt
Django==3.0.7
djangorestframework==3.11.0
Pillow==7.1.2
factory-boy==2.12.0